% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_do_okresu.R
\name{agreguj_do_okresu}
\alias{agreguj_do_okresu}
\title{agreguje dane do poziomu {id, id_zdau}}
\usage{
agreguj_do_okresu(dane)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
Agreguje dane do poziomu {id, id_zdau}
}
