% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_pkd.R
\name{agreguj_pkd}
\alias{agreguj_pkd}
\title{agreguje dane na poziomie skladek do poziomu {id, id_zdau, okres, pkd}}
\usage{
agreguj_pkd(skladki, okres, multidplyr = TRUE)
}
\arguments{
\item{skladki}{dane wygenerowane za pomocą funkcji
\code{\link{polacz_zus_zdau}}}

\item{okres}{okres agregacji (okienko agregacji liczone jest jako
\code{(okres - data_do) / okres})}

\item{multidplyr}{czy obliczać na wielu rdzeniach korzystając z pakietu
multidplyr}
}
\value{
data.frame wyliczone zmienne
}
\description{
Agreguje dane do poziomu {id, id_zdau, okres, pkd}
}
