% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bdl2teryt.R
\name{bdl2teryt}
\alias{bdl2teryt}
\title{Zamienia kody powiatów z kodów BDL na kody TERYT}
\usage{
bdl2teryt(id)
}
\arguments{
\item{id}{wektor kodów z BDL}
}
\value{
numeric kod TERYT powiatu
}
\description{
Zamienia kody powiatów z kodów BDL na kody TERYT
}

