% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data2okres.R
\name{data2okres}
\alias{data2okres}
\title{konwertuje datę w postaci tekstu/daty na liczbę miesięcy}
\usage{
data2okres(data)
}
\arguments{
\item{data}{data}
}
\value{
numeric liczba miesięcy jako ROK*12 + MIESIĄC
}
\description{
konwertuje datę w postaci tekstu/daty na liczbę miesięcy
}
\details{
Akceptowane są daty w formacie "{CYFRY}-{CYFRY}{COKOLWIEK}", gdzie pierwszy
ciąg cyfr interpretowany jest jako rok, a drugi jako liczba miesięcy. Stąd
funkcja działa poprawnie także dla niepełnych data (np. okres składkowy ZUS
"2014-05")
}
