% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_stale_czasowe.R
\name{konwertuj_daty}
\alias{konwertuj_daty}
\title{konwertuje istniejące stałe czasowe z powrotem na daty}
\usage{
konwertuj_daty(dane, data_badania)
}
\arguments{
\item{dane}{ramka danych ZDAU zwrócona przez funkcję \code{\link{przygotuj_zdau}}}

\item{data_badania}{data konca badanego okresu}
}
\value{
data.frame
}
\description{
konwertuje istniejące stałe czasowe z powrotem na daty
}
