% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/krok_zdau.R
\name{krok_zdau}
\alias{krok_zdau}
\title{wycina podzbiór obserwacji ze zbioru ZDAU}
\usage{
krok_zdau(zdau, n, tylkoAbs = TRUE)
}
\arguments{
\item{zdau}{ramka danych ZDAU zwrócona przez funkcję
\code{\link{przygotuj_zdau}}}

\item{n}{jaką liczbę rekordów ze zbioru zachować zaczynając od jakiego
rekordu, np. \code{n = c(0, 0)} zachowuje wszystkie rekordy, a \code{n =
c(1000, 15000)} zachowuje rekordy od 15001 do 16000}

\item{tylkoAbs}{czy pozostawić w danych ZDAU jedynie absolwentów}
}
\value{
data.frame z jedną zmienną - "id_zdau"
}
\description{
Typowy scenariusz użycia to wyliczanie zmiennych dla okresów w
środowisku o ograniczonym rozmiarze pamięci.

Funkcja jest stabilna, tzn. dla tych samych danych wejściowych zawsze zwraca
takie same dane.
}

