% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oblicz_absolwent_okres.R
\name{oblicz_absolwent_okres}
\alias{oblicz_absolwent_okres}
\title{oblicza zmienne generowane z poziomu (zus x zdau) przez poziom (zdau x okres) na poziom (zdau)}
\usage{
oblicz_absolwent_okres(dane)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne generowane z poziomu (zus x zdau) przez poziom (zdau x okres) na poziom (zdau)
}

