% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_etap.R
\name{oblicz_etap}
\alias{oblicz_etap}
\title{przypisuje odpowiedni etap kariery naukowej oraz jego długość}
\usage{
oblicz_etap(dane, kariera, dataMax)
}
\arguments{
\item{dane}{ramka danych zawierająca zmienną \code{data_od}}

\item{kariera}{dane wygenerowane za pomocą funkcji
\code{\link{oblicz_kariere}}}

\item{dataMax}{koniec badanego okresu (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
data.frame ramka danych z dołączonymi zmiennymi \code{etap} oraz
  \code{len}
}
\description{
przypisuje odpowiedni etap kariery naukowej oraz jego długość
}

