% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_kariere.R
\name{oblicz_kariere}
\alias{oblicz_kariere}
\title{oblicza przebieg kariery naukowej}
\usage{
oblicz_kariere(zdau, jednostki, kierunki)
}
\arguments{
\item{zdau}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zdau}}}

\item{jednostki}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_jednostki}}}
}
\value{
ramka danych opisująca karierę naukową (w postaci szerokiej)
}
\description{
oblicza przebieg kariery naukowej
}

