% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_nowi_pracodawcy.R
\name{oblicz_nowi_pracodawcy}
\alias{oblicz_nowi_pracodawcy}
\title{oblicza zmienne NNDN oraz NNNN}
\usage{
oblicz_nowi_pracodawcy(dane)
}
\arguments{
\item{dane}{zbiór danych wygenerowany funkcją \code{\link{oblicz_okienko}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne NNDN oraz NNNN
}

