% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oblicz_okienko.R
\name{oblicz_okienko}
\alias{oblicz_okienko}
\title{ustawia zmienne okienka czasowego (okres_min, okres_max, len)}
\usage{
oblicz_okienko(dane, okienkoMin, okienkoMax, dataMin, dataMax)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{polacz_zus_zdau}}
lub \code{\link{przygotuj_zdau}}}

\item{okienkoMin}{pierwszy miesiąc okienka względem daty uzyskania dyplomu}

\item{okienkoMax}{ostatni miesiąc okienka względem daty uzyskania dyplomu}

\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
data.frame wyliczone zmienne
}
\description{
ustawia zmienne okienka czasowego (okres_min, okres_max, len)
}

