% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_okienko_mld.R
\name{oblicz_okienko_mld}
\alias{oblicz_okienko_mld}
\title{wylicza zadane okienko danych w zakresie specyficznym dla monitorowania losów
doktorantów}
\usage{
oblicz_okienko_mld(okienko, zdau, publikacje, projekty, pomoc)
}
\arguments{
\item{okienko}{obiekt okienka utworzony za pomocą funkcji
\code{\link{okienko}}}

\item{zdau}{dane zwrócone przez funkcję \code{\link{przygotuj_zdau}}}

\item{publikacje}{dane zwrócone przez funkcję
\code{\link{przygotuj_publikacje}}}

\item{projekty}{dane miesieczne zwrócone przez funkcję
\code{\link{przygotuj_projekty}}}

\item{pomoc}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_pomoc}}}
}
\value{
[data.frame] dane obliczone dla wskazanego okienka czasu
}
\description{
wylicza zadane okienko danych w zakresie specyficznym dla monitorowania losów
doktorantów
}

