% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_osoba.R
\name{oblicz_osoba}
\alias{oblicz_osoba}
\title{oblicza zmienne generowane z poziomu (zus x zdau) na poziom (id) (dla danego
okienka czasu)}
\usage{
oblicz_osoba(dane, multidplyr = TRUE)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}}

\item{multidplyr}{czy obliczać na wielu rdzeniach korzystając z pakietu multidplyr}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne generowane z poziomu (zus x zdau) na poziom (id) (dla danego
okienka czasu)
}

