% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_pkd.R
\name{oblicz_pkd}
\alias{oblicz_pkd}
\title{oblicza zmienne ilustrujące zatrudnienie w poszczególnych działach PKD}
\usage{
oblicz_pkd(zus)
}
\arguments{
\item{zus}{dane zwrócone przez funkcję \code{\link{przygotuj_zus}}}

\item{okienko}{obiekt okienka utworzony za pomocą funkcji
\code{\link{okienko}}}
}
\value{
[data.frame] obliczone dane
}
\description{
oblicza zmienne ilustrujące zatrudnienie w poszczególnych działach PKD
}
