% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_pracodawcy.R
\name{oblicz_pracodawcy}
\alias{oblicz_pracodawcy}
\title{oblicza zmienne związane z liczbą pracodawców}
\usage{
oblicz_pracodawcy(dane, multidplyr = TRUE)
}
\arguments{
\item{dane}{zbiór danych wygenerowany funkcją \code{\link{oblicz_okienko}}}

\item{multidplyr}{czy obliczać na wielu rdzeniach korzystając z pakietu multidplyr}
}
\value{
data.frame wyliczone zmienne
}
\description{
Zmienne opisujące pracodawców wymagają danych źródłowych na poziomie nie
wyższym niż {osoba x płatnik x okres}, przez co nie mogą być liczone z danych
zagregowanych do poziomu miesięcy w funkcji \code{\link{agreguj_do_okresu}} i
wymagają osobnej funkcji.
}
