% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_publikacje.R
\name{oblicz_publikacje}
\alias{oblicz_publikacje}
\title{oblicza zmienne opisujące pomoc materialną}
\usage{
oblicz_publikacje(okienko, zdau, multidplyr = TRUE)
}
\arguments{
\item{okienko}{dane wygenerowane za pomocą funkcji
\code{\link{oblicz_okienko}} na danych jednostkowych (tzn. danych
wygenerowanych wcześniej funkcją \code{\link{przygotuj_pomoc}})}

\item{zdau}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zdau}}}

\item{multidplyr}{czy obliczać na wielu rdzeniach korzystając z pakietu
multidplyr}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne opisujące pomoc materialną
}

