% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oblicz_stale_czasowe.R
\name{oblicz_stale_czasowe}
\alias{oblicz_stale_czasowe}
\title{oblicza stałe czasowe i konwertuje istniejące stałe czasowe z powrotem na daty}
\usage{
oblicz_stale_czasowe(dane)
}
\arguments{
\item{dane}{ramka danych ZDAU (lub pochodna)}
}
\value{
data.frame
}
\description{
oblicza stałe czasowe i konwertuje istniejące stałe czasowe z powrotem na daty
}

