% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_studyp.R
\name{oblicz_studyp}
\alias{oblicz_studyp}
\title{oblicza zmienne KONT oraz STUDYP*}
\usage{
oblicz_studyp(dane)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne KONT oraz STUDYP*
}
