% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generuj_wartosci.R
\name{oblicz_wartosci}
\alias{oblicz_wartosci}
\title{generuje wartości zadanych zmiennych dla zadanych odbiorców}
\usage{
oblicz_wartosci(odbiorcy, zbDanych, zmienne, grupa, prefiks, stale = 1:9,
  pomin = "NIE DOTYCZY", nMin = 10)
}
\arguments{
\item{odbiorcy}{ramka danych opisująca odbiorców}

\item{zbDanych}{lista zawierająca zbiory danych na podstawie których
wyliczane będą wartości zmiennych}

\item{zmienne}{wektor definicji zmiennych (wektor stringów, patrz też
parametr \code{pomin})}

\item{grupa}{warunek wybierający obserwacje należące do danego odbiorcy (w
postaci stringu)}

\item{prefiks}{prefiks nazw generowanych zmiennych}

\item{stale}{wektor indeksów parametru \code{zmienne} wskazujący zmienne,
które nie podlegają anonimizacji (są generowane bez względu na liczebność
grupy - patrz parametr \code{nMin})}

\item{pomin}{wektor wartości oznaczających, że dana zmienna powinna zostać
pominięta (patrz sekcja "details")}

\item{nMin}{liczba obserwacji poniżej której wartości nie są wyliczane}
}
\value{
data.frame ramka danych z wyliczonymi wartościami zmiennych
}
\description{
Raporty w postaci plików pdf są trudne do automatycznego
przetwarzania, stąd potrzeba wygenerowania zawartych w nich informacji w
formie "zwyczajnych" zbiorów danych, co umożliwia niniejsza funkcja.

Niestety w chwili obecnej zamiast parsować szablon raportu wymaga manualnego
podania definicji wszystkich zmiennych.
}
\details{
Zmienne w zwracanej ramce danych noszą nazwy będące połączeniem parametru
\code{prefiks} oraz ich indeksu w parametrze \code{zmienne}. Jeśli dana
zmienna została pominięta (patrz parametr \code{pomin}), nie powoduje to
przenumerowania zmiennych następujących za nią (np. gdy \code{zmienne =
c('formuła1', 'NIE DOTYCZY', 'formuła2')} i \code{prefiks = 'p_'}, wtedy
wygenerowane zostaną zmienne \code{p_1} oraz \code{p_3}).
}
