% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oblicz_zamieszkanie.R
\name{oblicz_zamieszkanie}
\alias{oblicz_zamieszkanie}
\title{oblicza zmienne związane z miejscem zamieszkania}
\usage{
oblicz_zamieszkanie(dane, jednostki, wMomDyplomu)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}}

\item{jednostki}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_jednostki}}}

\item{wMomDyplomu}{czy wyliczyć w momencie dyplomu czy dla końca okienka}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne związane z miejscem zamieszkania
}

