% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_zmienne_czasowe.R
\name{oblicz_zmienne_czasowe}
\alias{oblicz_zmienne_czasowe}
\title{oblicza zmienne związane z czasem, który upłynął od momentu uzyskania dyplomu}
\usage{
oblicz_zmienne_czasowe(dane, utrataPracy)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{polacz_zus_zdau}}}

\item{utrataPracy}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_utrata_pracy}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne związane z czasem, który upłynął od momentu uzyskania dyplomu
}
