% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_zmienne_miesieczne.R
\name{oblicz_zmienne_miesieczne}
\alias{oblicz_zmienne_miesieczne}
\title{oblicza zmienne generowane z poziomu (zus x zdau) przez poziom (zdau x okres) na poziom (zdau)}
\usage{
oblicz_zmienne_miesieczne(dane, zdau, multidplyr = TRUE)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}}

\item{multidplyr}{czy obliczać na wielu rdzeniach korzystając z pakietu multidplyr}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne generowane z poziomu (zus x zdau) przez poziom (zdau x okres) na poziom (zdau)
}

