% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okienko.R
\name{okienko}
\alias{okienko}
\title{tworzy obiekt opisujący okienko czasu}
\usage{
okienko(offsetMin, offsetMax, zmiennaMin, zmiennaMax, sufiks, dataMin,
  dataMax)
}
\arguments{
\item{offsetMin}{pierwszy miesiąc okienka względem wartości zmiennej
wskazywanej przez parametr \code{zmiennaMin}}

\item{offsetMax}{ostatni miesiąc okienka względem wartości zmiennej
wskazywanej przez parametr \code{zmiennaMax}}

\item{zmiennaMin}{nazwa zmiennej, względem której obliczany jest początek
okienka}

\item{zmiennaMax}{nazwa zmiennej, względem której obliczany jest koniec
okienka}

\item{sufiks}{sufiks nazw zmiennych tego okienka czasu}

\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2015-09-30')}
}
\value{
[okienko] obiekt opisujący okienko czasu
}
\description{
tworzy obiekt opisujący okienko czasu
}
