% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okienko_ela.R
\name{okienko_ela}
\alias{okienko_ela}
\title{wylicza zadane okienka danych zgodnie z potrzebami projektu ELA}
\usage{
okienko_ela(okienko, zdau, baza, miesieczne, utrataPracy, jednostki,
  krok = 1e+06, info = TRUE)
}
\arguments{
\item{okienko}{obiekt okienka utworzony za pomocą funkcji
\code{\link{okienko}}}

\item{zdau}{dane zwrócone przez funkcję \code{\link{przygotuj_zdau}}}

\item{baza}{dane zwrócone przez funkcję \code{\link{polacz_zus_zdau}}}

\item{miesieczne}{dane miesieczne zwrócone przez funkcję
\code{\link{agreguj_do_miesiecy}}}

\item{utrataPracy}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_utrata_pracy}}}

\item{jednostki}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_jednostki}}}

\item{krok}{liczba obserwacji obliczanych w pojedynczym kroku (determinuje
rozmiar potrzebnej pamięci - im większa, tym więcej pamięci potrzeba, ale
też obliczenia przebiegają nieco szybciej)}

\item{info}{czy wyświetlać informacje o przetwarzanym okienku}
}
\value{
[data.frame] dane obliczone dla wskazanego okienka czasu
}
\description{
wylicza zadane okienka danych zgodnie z potrzebami projektu ELA
}
