% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/okres2data.R
\name{okres2data}
\alias{okres2data}
\title{konwertuje okres składkowy (rok i miesiąc) wyrażony liczbą na datę}
\usage{
okres2data(okres)
}
\arguments{
\item{okres}{liczba miesięcy wyrażona liczbą}
}
\value{
character data wyrażona jako łańcuch znaków (np. '2004-11')
}
\description{
konwertuje okres składkowy (rok i miesiąc) wyrażony liczbą na datę
}

