% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/okres2miesiac.R
\name{okres2miesiac}
\alias{okres2miesiac}
\title{odczytuje miesiąc z liczby miesięcy wyrażonej liczbą}
\usage{
okres2miesiac(okres)
}
\arguments{
\item{okres}{liczba miesięcy wyrażona liczbą}
}
\value{
numeric miesiąc (1-12)
}
\description{
odczytuje miesiąc z liczby miesięcy wyrażonej liczbą
}
\seealso{
data2okres
}

