% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/okres2rok.R
\name{okres2rok}
\alias{okres2rok}
\title{odczytuje rok z liczby miesięcy wyrażonej liczbą}
\usage{
okres2rok(okres)
}
\arguments{
\item{okres}{liczba miesięcy wyrażona liczbą}
}
\value{
numeric rok
}
\description{
odczytuje rok z liczby miesięcy wyrażonej liczbą
}
\seealso{
data2okres
}

