% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pokaz_roznice.R
\name{pokaz_roznice}
\alias{pokaz_roznice}
\title{Odfiltrowuje rekordy z różnicami wartości wskazanej zmiennej}
\usage{
pokaz_roznice(dane, zmienna, zmienneId = c("id_zdau", "id", "ID_ZDAU",
  "ID", "okres", "OKRES", "id_zdu3"), tolerancja = 1e-05)
}
\arguments{
\item{dane}{zbiór danych}

\item{zmienna}{prefiks zmiennej, między którymi występują różnice}

\item{zmienneId}{list zmiennych, która powinna zostać dołączona do wyniku}

\item{tolerancja}{tolerancja przy porównywaniu liczb zmiennoprzecinkowych}
}
\value{
data.frame odfiltrowane dane
}
\description{
Używać na wyniku funkcji \code{\link{porownaj_zbiory}}
}
