% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polacz_pna_powiaty.R
\name{polacz_pna_powiaty}
\alias{polacz_pna_powiaty}
\title{przyłącza dane powiatów do danych PNA i agreguje je tak, by (rok, pna) było
unikalne}
\usage{
polacz_pna_powiaty(pna, powiaty, dataMin, dataMax)
}
\arguments{
\item{pna}{dane PNA wygenerowane za pomocą funkcji \code{\link{przygotuj_pna}}}

\item{powiaty}{dane powiatów wygenerowane za pomocą funkcji \code{\link{przygotuj_powiaty}}}

\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
data.frame wyliczone dane
}
\description{
przyłącza dane powiatów do danych PNA i agreguje je tak, by (rok, pna) było
unikalne
}

