% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polacz_w_baze.R
\name{polacz_w_baze}
\alias{polacz_w_baze}
\title{łączy dane zus, zdau, statystyki powiatów, informacje o jednostkach oraz
karierze}
\usage{
polacz_w_baze(zus, zdau, kariera, jednostki, pnaPowiaty, dataMin, dataMax)
}
\arguments{
\item{zus}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zus}}}

\item{zdau}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zdau}}}

\item{kariera}{dane wygenerowane za pomocą funkcji
\code{\link{oblicz_kariere}}}

\item{jednostki}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_jednostki}}}

\item{pnaPowiaty}{dane wygenerowane za pomocą funkcji
\code{\link{polacz_pna_powiaty}}}

\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2015-09-30')}
}
\value{
data.frame wyliczone zmienne
}
\description{
Aby poprawnie przyłączyć statystyki powiatów macierz danych
uzupełniana jest tak, aby dla każdego absolwenta zawierała informacje o
każdym okresie w zadanym przedziale czasu, ale nie później niż do "końca
obserwacji w zus" (zakładając zamieszkanie absolwenta w uzupełnianych
okresach "w Polsce", kodowane jako pna = -1, oraz wszelkie cechy tytułu
ubezpieczenia ZUS równe 0, w szczególności także bezrob = 0).

Generowana jest również zmienna określająca, czy absolwent studiował w
poszczególnych okresach, wyznaczana na podstawie danych ZDAU
}

