% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polacz_zus_zdau.R
\name{polacz_zus_zdau}
\alias{polacz_zus_zdau}
\title{łączy dane zus i zdau}
\usage{
polacz_zus_zdau(zus, zdau, pnaPowiaty, dataMin, dataMax)
}
\arguments{
\item{zus}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zus}}}

\item{zdau}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zdau}}}

\item{pnaPowiaty}{dane wygenerowane za pomocą funkcji \code{\link{polacz_pna_powiaty}}}

\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
data.frame wyliczone zmienne
}
\description{
łączy dane zus i zdau
}

