% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/porownaj.R
\name{porownaj}
\alias{porownaj}
\title{zestawia ze sobą wartości wskazanej zmiennej w dwóch zbiorach}
\usage{
porownaj(a, b, zmiennaA, zmiennaB = zmiennaA, id = c("ID_ZDAU", "ID"))
}
\arguments{
\item{a}{pierwszy zbiór}

\item{b}{drugi zbiór}

\item{zmiennaA}{nazwa zmiennej do porównania w zbiorze a}

\item{zmiennaB}{nazwa zmiennej do porównania w zbiorze b}

\item{id}{wektor nazw zmiennych stanowiących klucz podstawowy w zbiorach}
}
\value{
data.frame wartości zmiennej połączone z obydwy zbiorów
}
\description{
zestawia ze sobą wartości wskazanej zmiennej w dwóch zbiorach
}

