% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_przygotuj_zus_zdau.R
\name{pre_przygotuj_zus_zdau}
\alias{pre_przygotuj_zus_zdau}
\title{Złącza dane ZUS z eskportów dla wielu roczników absolwentów w jeden zestaw
danych}
\usage{
pre_przygotuj_zus_zdau(katalog)
}
\arguments{
\item{katalog}{lokalizacja zawierająca katalogi z danymi dla poszczególnych
roczników absolwentów (ich nazwy powinny sortować się w takiej kolejności,
jak roczniki absolwentów)}
}
\description{
OPI eksportuje oddzielny zbiór ZDAU dla każdego rocznika
absolwentów uwzględniając w nim informacje o innych kierunkach studiów
jedynie jeśli okres studiowania przypadał na ten sam rok, co rok uzyskania
dyplomu. Uniemożliwia to poprawne śledzenie kontynuacji nauki na innych
kierunkach studiów. Drugą niedogodnością jest wielokrotne przetwarzanie
informacji z ZUS dotyczących tych samych osób będących absolwentami (różnych
kierunków) w różnych rocznikach.

Aby rozwiazać obydwie niedogodności należy złączyć zbiory opisujące
poszczególne roczniki absolwentów w jeden zestaw danych, do czego służy ta
funkcja.
}
