% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_jednostki_old.R
\name{przygotuj_jednostki_old}
\alias{przygotuj_jednostki_old}
\title{Przygotowuje słownik uczelni}
\usage{
przygotuj_jednostki_old(katZr, rok)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik sl_instytucje.xlsx}

\item{rok}{rok, który zostanie przypisany jednostkom (gdy NA, obecny rok)}
}
\value{
[data.frame] ramka danych opisująca uczelnie
}
\description{
Zwraca słownik uczelni
}
