% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_kierunki.R
\name{przygotuj_kierunki}
\alias{przygotuj_kierunki}
\title{Przygotowuje słownik kierunków}
\usage{
przygotuj_kierunki(katZr, rok)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik sl_instytucje.xlsx}

\item{rok}{rok, który zostanie przypisany jednostkom}
}
\value{
[data.frame] ramka danych opisująca kierunki studiów
}
\description{
Zwraca słownik kierunków studiów

Uwaga, kluczem unikalnym zwracanych danych jest para {kierunek_id,
jednostka_id}!
}

