% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_kierunki_old.R
\name{przygotuj_kierunki_old}
\alias{przygotuj_kierunki_old}
\title{Przygotowuje słownik kierunków}
\usage{
przygotuj_kierunki_old(katZr, zmDodatkowe = character())
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik sl_instytucje.xlsx}

\item{zmDodatkowe}{nazwy dodatkowych zmiennych do pozostawienia w zwracanym zbiorze danych}
}
\value{
[data.frame] ramka danych opisująca kierunki studiów
}
\description{
Zwraca słownik kierunków studiów
}
