% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_pna_powiaty.R
\name{przygotuj_pna_powiaty}
\alias{przygotuj_pna_powiaty}
\title{przygotowuje dane o PNA i powiatach}
\usage{
przygotuj_pna_powiaty(dataMin, dataMax)
}
\arguments{
\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
data.frame wyliczone dane
}
\description{
przygotowuje dane o PNA i powiatach
}
