% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/przygotuj_pna_powiaty_mb.R
\name{przygotuj_pna_powiaty_mb}
\alias{przygotuj_pna_powiaty_mb}
\title{przygotowuje dane o PNA i powiatach na podstawie zbioru danych od Marka
Bożykowskiego}
\usage{
przygotuj_pna_powiaty_mb(dataMin, dataMax)
}
\arguments{
\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
data.frame wyliczone dane
}
\description{
przygotowuje dane o PNA i powiatach na podstawie zbioru danych od Marka
Bożykowskiego
}

