% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_pomoc.R
\name{przygotuj_pomoc}
\alias{przygotuj_pomoc}
\title{Przygotowuje zbiór z informacjami o pomocy materialnej}
\usage{
przygotuj_pomoc(katZr, kariera, dataMax)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik pomoc.csv}

\item{kariera}{dane wygenerowane za pomocą funkcji
\code{\link{oblicz_kariere}}}

\item{dataMax}{koniec badanego okresu (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
ramka danych z informacjami o pomocy materialnej
}
\description{
Przygotowuje zbiór z informacjami o pomocy materialnej
}

