% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_powiaty.R
\name{przygotuj_powiaty}
\alias{przygotuj_powiaty}
\title{przygotowuje powiatowe dane o bezrobociu i zarobkach}
\usage{
przygotuj_powiaty()
}
\description{
przygotowuje powiatowe dane o bezrobociu i zarobkach
}
