% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_publikacje.R
\name{przygotuj_publikacje}
\alias{przygotuj_publikacje}
\title{Przygotowuje zbiór z informacjami o publikacjach naukowych}
\usage{
przygotuj_publikacje(katZr, kariera)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik publikacje.csv}

\item{kariera}{dane wygenerowane za pomocą funkcji
\code{\link{oblicz_kariere}}}
}
\value{
ramka danych z informacjami o publikacjach naukowych
}
\description{
Przygotowuje zbiór z informacjami o publikacjach naukowych
}

