% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_uczelnie.R
\name{przygotuj_uczelnie}
\alias{przygotuj_uczelnie}
\title{Przygotowuje słownik uczelni}
\usage{
przygotuj_uczelnie(katZr)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik sl_instytucje.xlsx}
}
\value{
[data.frame] ramka danych opisująca uczelnie
}
\description{
Zwraca słownik uczelni
}
