% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_utrata_pracy.R
\name{przygotuj_utrata_pracy}
\alias{przygotuj_utrata_pracy}
\title{oblicza zmienne utraty pracy, używane do wyliczania zmiennych NMLE, NMLEP
oraz zmiennych czasowych}
\usage{
przygotuj_utrata_pracy(zus, dataMax, multidplyr = TRUE)
}
\arguments{
\item{zus}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zus}}}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch
znaków, np. '2015-09-30')}

\item{multidplyr}{czy obliczać na wielu rdzeniach korzystając z pakietu
multidplyr}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne utraty pracy, używane do wyliczania zmiennych NMLE, NMLEP
oraz zmiennych czasowych
}
