% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_zdu1.R
\name{przygotuj_zdu1}
\alias{przygotuj_zdu1}
\title{Wczytuje zbiór ZDU1}
\usage{
przygotuj_zdu1(katZr)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik ZDU1.csv}
}
\description{
Wczytuje zbiór opisujący osoby ubezpieczone (ZDU1)
}
