% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_zus.R
\name{przygotuj_zus}
\alias{przygotuj_zus}
\title{Przygotowuje dane z ZUS}
\usage{
przygotuj_zus(katZr, dataMin, dataMax, pna, polSparka, cacheTestow = TRUE)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik ZDUx.csv}

\item{dataMin}{pierwszy uwzględniany okres składkowy}

\item{dataMax}{ostatni uwzględniany okres składkowy}

\item{pna}{ramka danych z dozwolonymi kodami pna}

\item{polSparka}{połączenie ze Sparkiem}

\item{cacheTestow}{czy cache-ować dane przed testami poprawności (jeśli tylko
dane mieszczą się w RAM-ie, znacznie przyspiesza wykonywani testów
poprawności)}
}
\description{
Do danych o składkach dołączane są dane o miejscu zamieszkania, płatniku,
tytułach ubezpieczeń, itp.
}
