% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_csv_spark.R
\name{read_csv_spark}
\alias{read_csv_spark}
\title{wczytuje dane zapisane za pomocą sparklyr::spark_write_csv()}
\usage{
read_csv_spark(sciezka, typyKolumn = NULL)
}
\arguments{
\item{sciezka}{ścieżka do pliku/katalogu przechowującego dane}

\item{typyKolumn}{opcjonalnie typy kolumn w formacie parametru
\code{col_types} funkcji \code{\link[readr]{read_csv}} (d - double, i -
integer, c - character)}
}
\value{
data.frame
}
\description{
sparklyr::spark_write_csv() zapisuje większe zbiory danych w częściach.
Ta funkcja umożliwia ich łatwe wczytywanie
}
