% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_col_types.R
\name{tbl_col_types}
\alias{tbl_col_types}
\title{Zwraca opis typów danych w formacie readr::read_csv()}
\usage{
tbl_col_types(tbl)
}
\arguments{
\item{tbl}{ramka danych korzystająca z dowolnego backendu}
}
\value{
character
}
\description{
Zwraca opis typów danych w formacie readr::read_csv()
}
