% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ustaw_okienko_etap.R
\name{ustaw_okienko_etap}
\alias{ustaw_okienko_etap}
\title{Odfiltrowuje obserwacje na podstawie zmiennej \code{etap}}
\usage{
ustaw_okienko_etap(dane, okienko)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_publikacje}}, \code{\link{przygotuj_pomoc}} lub
\code{\link{przygotuj_projekty}}}

\item{okienko}{obiekt opisujący okienko stworzony za pomocą funkcji
\code{\link{okienko}}}
}
\value{
data.frame odfiltrowane dane
}
\description{
Wartość zmiennej etap do odfiltrowania pobierana jest z sufiksu okienka
poprzez usunięcie znaków podkreślenia i zamianę pisowni na wielkie litery.
}

