% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uzgodnij_teryt.R
\name{uzgodnij_teryt}
\alias{uzgodnij_teryt}
\title{uzgadnia TERYT}
\usage{
uzgodnij_teryt(teryt)
}
\arguments{
\item{teryt}{wektor kodów TERYT}
}
\value{
numeric uzgodniony kod TERYT
}
\description{
W razie niezgodności zwraca najbliższą wspólną JST (powiat, województwo bądz
Polskę)
}

