% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uzupelnij_obserwacje.R
\name{uzupelnij_obserwacje}
\alias{uzupelnij_obserwacje}
\title{uzupełnia rekordy ramki danych o unikalne obserwacje z innej ramki danych}
\usage{
uzupelnij_obserwacje(wynik, wzor, id = "id_zdau")
}
\arguments{
\item{wynik}{ramka danych do uzupełnienia}

\item{wzor}{ramka danych zawierająca wszystkie obserwacje}

\item{id}{wektor nazw zmiennych wyznaczających unikalne obserwacje w ramce
danych \code{wzor}}
}
\value{
data.frame dane z parametru \code{wynik} uzupełnione o wszystkie
  unikalne identyfikatory z parametru \code{wzor}
}
\description{
uzupełnia rekordy ramki danych o unikalne obserwacje z innej ramki danych
}

