% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapisz_dla_sparka.R
\name{zapisz_dla_sparka}
\alias{zapisz_dla_sparka}
\title{Zapisuje ramkę danych w formacie CSV kompatybilnym ze Sparkiem}
\usage{
zapisz_dla_sparka(dane, nazwaPliku, katalog = NULL)
}
\arguments{
\item{dane}{ramka danych do zapisania}

\item{nazwaPliku}{nazwa pliku, w którym zpaisane zostaną dane (bez
rozszerzenia)}

\item{katalog}{katalog, w którym zapisane zostaną dane, jeśli nie podany,
będzie to podkatalog "cache" w katalogu roboczym}
}
\value{
character ścieżka zapisu pliku
}
\description{
Zapisuje ramkę danych w formacie CSV kompatybilnym ze Sparkiem
}
