% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zapisz_ze_sparka.R
\name{zapisz_ze_sparka}
\alias{zapisz_ze_sparka}
\title{Eksportuje dane ze Sparka do CSV}
\usage{
zapisz_ze_sparka(dane, nazwaPliku, buforuj = FALSE, katalog = NULL,
  polSparka = NULL, zlacz = TRUE)
}
\arguments{
\item{dane}{ramka danych Sparka}

\item{nazwaPliku}{nazwa pliku CSV (bez rozszerzenia)}

\item{buforuj}{czy zbuforować ramkę danych przed zapisem}

\item{katalog}{katalog, w którym zapisane zostaną dane, jeśli nie podany,
będzie to podkatalog "cache" w katalogu roboczym}

\item{polSparka}{połączeni ze Sparkiem, niezbędne gdy \code{buforuj = TRUE}}

\item{zlacz}{czy złączyć wszystkie partycje Sparka w jeden zbiór}
}
\value{
character ścieżka zapisu
}
\description{
Eksportuje dane ze Sparka do CSV
}
