% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agreguj_do_miesiecy.R
\name{agreguj_do_miesiecy}
\alias{agreguj_do_miesiecy}
\title{agreguje dane do poziomu {id, id_zdau, okres}}
\usage{
agreguj_do_miesiecy(dane, zdau)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{polacz_zus_zdau}}}

\item{zdau}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zdau}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
Agreguje dane do poziomu {id, id_zdau, okres}
}
