% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nazwa_pliku.R
\name{nazwa_pliku}
\alias{nazwa_pliku}
\title{generuje nazwę pliku zgodną z konwencją}
\usage{
nazwa_pliku(zbior, rozszerzenie = ".RData", katalog = "cache",
  rocznik = "")
}
\arguments{
\item{zbior}{nazwa zbioru}

\item{rozszerzenie}{rozszerzenie pliku (domyślnie .RData)}

\item{katalog}{ścieżka do katalogu}

\item{rocznik}{rocznik absolwentów (np. 2014)}
}
\description{
Ułatwia generowanie nazw plików w powtarzalny sposób
}
