% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oblicz_absolwent.R
\name{oblicz_absolwent}
\alias{oblicz_absolwent}
\title{oblicza zmienne generowane w jednym kroku z poziomu (zus x zdau) na poziom (zdau)}
\usage{
oblicz_absolwent(dane)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne generowane w jednym kroku z poziomu (zus x zdau) na poziom (zdau)
}

