% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_okienko.R
\name{oblicz_okienko}
\alias{oblicz_okienko}
\title{ustawia zmienne okienka czasowego (okres_min, okres_max, len)}
\usage{
oblicz_okienko(dane, okienko, filtrZdau = NULL)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{polacz_zus_zdau}},
lub \code{\link{agreguj_do_miesiecy}}}

\item{okienko}{obiekt opisujący okienko stworzony za pomocą funkcji
\code{\link{okienko}}}

\item{filtrZdau}{ramka danych zawierająca zmienną \code{id_zdau} ograniczająca
obserwacje, które mają się znaleźć w okienku}
}
\value{
data.frame wyliczone zmienne
}
\description{
ustawia zmienne okienka czasowego (okres_min, okres_max, len)
}

