% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_stale.R
\name{oblicz_stale}
\alias{oblicz_stale}
\title{oblicza proste zmienne niezwiazane z okienkami czasu}
\usage{
oblicz_stale(dane, zdau)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{polacz_zus_zdau}}}

\item{zdau}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zdau}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza proste zmienne niezwiazane z okienkami czasu
}
