% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_stale_czasowe.R
\name{oblicz_stale_czasowe}
\alias{oblicz_stale_czasowe}
\title{oblicza stałe czasowe i konwertuje istniejące stałe czasowe z powrotem na daty}
\usage{
oblicz_stale_czasowe(dane, data_badania)
}
\arguments{
\item{dane}{ramka danych ZDAU zwrócona przez funkcję \code{\link{przygotuj_zdau}}}

\item{data_badania}{data konca badanego okresu}
}
\value{
data.frame
}
\description{
oblicza stałe czasowe i konwertuje istniejące stałe czasowe z powrotem na daty
}
