% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oblicz_zamieszkanie.R
\name{oblicz_zamieszkanie}
\alias{oblicz_zamieszkanie}
\title{oblicza zmienne związane z miejscem zamieszkania}
\usage{
oblicz_zamieszkanie(dane, jednostki, multidplyr = FALSE)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}
z danych wygenerowanych wcześniej funkcją \code{\link{polacz_w_baze}})}

\item{jednostki}{dane wygenerowane za pomocą funkcji
\code{\link{przygotuj_jednostki}}}

\item{multidplyr}{czy obliczać na wielu rdzeniach korzystając z pakietu
multidplyr}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne związane z miejscem zamieszkania
}

