% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oblicz_zmienne_czasowe.R
\name{oblicz_zmienne_czasowe}
\alias{oblicz_zmienne_czasowe}
\title{oblicza zmienne związane z czasem, który upłynął od momentu uzyskania dyplomu}
\usage{
oblicz_zmienne_czasowe(dane, utrataEtatu)
}
\arguments{
\item{dane}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_okienko}}}

\item{utrataEtatu}{dane wygenerowane za pomocą funkcji \code{\link{oblicz_utrata_etatu}}}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne związane z czasem, który upłynął od momentu uzyskania dyplomu
}

