% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/oblicz_zmienne_pochodne.R
\name{oblicz_zmienne_pochodne}
\alias{oblicz_zmienne_pochodne}
\title{oblicza zmienne wywodzone z wcześniej policzonych zmiennych}
\usage{
oblicz_zmienne_pochodne(dane)
}
\arguments{
\item{dane}{połączone dane na poziomie (zdau) wygenerowane za pomocą innych
funkcji}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne wywodzone z wcześniej policzonych zmiennych
}

