% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/polacz_zus_zdau.R
\name{polacz_zus_zdau}
\alias{polacz_zus_zdau}
\title{łączy dane zus, zdau oraz statystyki powiatów}
\usage{
polacz_zus_zdau(zus, zdau, pnaPowiaty, dataMin, dataMax)
}
\arguments{
\item{zus}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zus}}}

\item{zdau}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zdau}}}

\item{pnaPowiaty}{dane wygenerowane za pomocą funkcji \code{\link{polacz_pna_powiaty}}}

\item{dataMin}{początek okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2014-01-01')}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
data.frame wyliczone zmienne
}
\description{
Aby poprawnie przyłączyć statystyki powiatów macierz danych uzupełniana jest
tak, aby dla każdego absolwenta zawierała informacje o każdym okrsie w
zadanym przedziale czasu (zakładając zamieszkanie absolwenta w uzupełnianych
okresach "w Polsce", kodowane jako pna = -1, oraz wszelkie cechy tytułu
ubezpieczenia ZUS równe 0, w szczególności także bezrob = 0).
}

