% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_jednostki.R
\name{przygotuj_jednostki}
\alias{przygotuj_jednostki}
\title{Przygotowuje słownik jednostek}
\usage{
przygotuj_jednostki(rok = NA)
}
\arguments{
\item{rok}{rok, który zostanie przypisany jednostkom (gdy NA, obecny rok)}
}
\description{
Przygotowuje słownik jednostek
}

