% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_jednostki.R
\name{przygotuj_jednostki}
\alias{przygotuj_jednostki}
\title{Przygotowuje słownik jednostek}
\usage{
przygotuj_jednostki(katZr, zmiennaRok = "rok")
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik sl_instytucje.xlsx}

\item{zmiennaRok}{nazwa zmiennej z rokiem rekordu}
}
\value{
[data.frame] ramka danych opisująca jednostki
}
\description{
Zwraca słownik jednostek
}
