% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_kierunki.R
\name{przygotuj_kierunki}
\alias{przygotuj_kierunki}
\title{Przygotowuje słownik kierunków}
\usage{
przygotuj_kierunki(katZr, agregujDoKierunku = TRUE, jednostki = NULL,
  zmiennaRok = "rok")
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik KIERUNKI.xlsx}

\item{agregujDoKierunku}{czy agregować dane tak by {rok, kierunek_id} były
unikalne (jeśli FALSE, agregacja nastapi do unikalności na poziomie {rok,
kierunek_id, jednostka_id})}

\item{jednostki}{ramka danych opisująca jednostki (gdy NULL, zostanie
wczytana automatycznie)}

\item{zmiennaRok}{nazwa zmiennej z rokiem rekordu}
}
\value{
[data.frame] ramka danych opisująca kierunki studiów
}
\description{
Zwraca słownik kierunków studiów
}
