% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_projekty.R
\name{przygotuj_projekty}
\alias{przygotuj_projekty}
\title{Przygotowuje zbiór z informacjami o projektach naukowych}
\usage{
przygotuj_projekty(katZr, kariera, dataMax)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik projekty.csv}

\item{kariera}{dane wygenerowane za pomocą funkcji
\code{\link{oblicz_kariere}}}

\item{dataMax}{koniec badanego okresu (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
ramka danych z informacjami o projektach naukowych
}
\description{
Przygotowuje zbiór z informacjami o projektach naukowych
}

