% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/przygotuj_utrata_pracy.R
\name{przygotuj_utrata_pracy}
\alias{przygotuj_utrata_pracy}
\title{oblicza zmienne utraty pracy, używane do wyliczania zmiennych NMLE oraz NMLEP}
\usage{
przygotuj_utrata_pracy(zus, dataMax)
}
\arguments{
\item{zus}{dane wygenerowane za pomocą funkcji \code{\link{przygotuj_zus}}}

\item{dataMax}{koniec okresu uwzględnionego w danych ZUS (jako łańcuch znaków, np. '2015-09-30')}
}
\value{
data.frame wyliczone zmienne
}
\description{
oblicza zmienne utraty pracy, używane do wyliczania zmiennych NMLE oraz NMLEP
}

