% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/przygotuj_zus.R
\name{przygotuj_zus}
\alias{przygotuj_zus}
\title{Przygotowuje dane z ZUS}
\usage{
przygotuj_zus(dataMin, dataMax)
}
\arguments{
\item{dataMin}{pierwszy uwzględniany okres składkowy}

\item{dataMax}{ostatni uwzględniany okres składkowy}
}
\description{
Do danych o składkach dołączane są dane o miejscu zamieszkania, płatniku,
tytułach ubezpieczeń, itp.
}

