% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/przygotuj_zus.R
\name{przygotuj_zus}
\alias{przygotuj_zus}
\title{Przygotowuje dane z ZUS}
\usage{
przygotuj_zus(katZr, dataMin, dataMax, pna, multidplyr = TRUE)
}
\arguments{
\item{katZr}{katalog, w którym znajduje się plik ZDUx.csv}

\item{dataMin}{pierwszy uwzględniany okres składkowy}

\item{dataMax}{ostatni uwzględniany okres składkowy}

\item{pna}{ramka danych z dozwolonymi kodami pna}

\item{multidplyr}{czy obliczać na wielu rdzeniach korzystając z pakietu
multidplyr}
}
\description{
Do danych o składkach dołączane są dane o miejscu zamieszkania, płatniku,
tytułach ubezpieczeń, itp.
}

